fix: extract shared E2E helpers, fix stale form selectors in admin/team specs
admin.spec.ts and team.spec.ts still had the old createHousehold helper
(missing 'Confirm Admin PIN' field, wrong label selectors) causing all
tests to time out on waitForURL. Extracted createHousehold/fillPinInput
into e2e/helpers.ts so all specs stay in sync.
