+++ nestle-together/e2e/team.spec.ts
@@ -1,23 +1,8 @@
 import { test, expect } from '@playwright/test';
-
-const uniqueId = () => Math.random().toString(36).substring(7);
+import { createHousehold, fillPinInput, uniqueId } from './helpers';
 
 const ADMIN_PIN = '1234';
 const MEMBER_PIN = '5678';
-
-async function createHousehold(page: import('@playwright/test').Page) {
-  const householdName = `Team Test ${uniqueId()}`;
-  await page.goto('/');
-  await page.click('text=Create Household');
-  await page.getByLabel(/household name/i).fill(householdName);
-  await page.getByLabel(/your name|nickname/i).first().fill('Admin');
-  await page.getByRole('button', { name: /continue/i }).click();
-  await page.getByLabel(/admin.*pin|pin.*code/i).first().fill(ADMIN_PIN);
-  await page.getByRole('button', { name: /create/i }).click();
-  await page.waitForURL(/\/household\//, { timeout: 45000 });
-  await expect(page.locator('h1')).toContainText(householdName, { timeout: 10000 });
-  return page.url();
-}
 
 async function setMemberPin(page: import('@playwright/test').Page) {
   // Navigate to Admin → Settings tab to set member PIN
@@ -37,7 +22,7 @@ async function navigateToTeamTab(page: import('@playwright/test').Page) {
 
 test.describe('Team tab — admin view', () => {
   test('shows Household Overview with reassignment hint', async ({ page }) => {
-    await createHousehold(page);
+    await createHousehold(page, `Team Test ${uniqueId()}`);
     await navigateToTeamTab(page);
 
     await expect(page.locator('text=Household Overview')).toBeVisible();
@@ -45,7 +30,7 @@ test.describe('Team tab — admin view', () => {
   });
 
   test('shows gear icon when chore is assigned', async ({ page }) => {
-    await createHousehold(page);
+    await createHousehold(page, `Team Test ${uniqueId()}`);
 
     // Add a chore via Admin → Chores tab
     await page.getByRole('link', { name: /admin/i }).click();
@@ -69,17 +54,16 @@ test.describe('Team tab — member view', () => {
   let householdAccessUrl: string;
 
   test.beforeEach(async ({ page }) => {
-    const dashboardUrl = await createHousehold(page);
+    const dashboardUrl = await createHousehold(page, `Team Test ${uniqueId()}`, ADMIN_PIN);
     await setMemberPin(page);
 
     // Logout and re-access with member PIN
     householdAccessUrl = dashboardUrl.replace('/household/', '/access/');
-    await page.getByRole('button', { name: /logout|log out/i }).click();
+    await page.getByRole('button', { name: /log out/i }).click();
 
     await page.goto(householdAccessUrl);
-    await page.getByLabel(/pin/i).fill(MEMBER_PIN);
-    await page.getByRole('button', { name: /access|enter|submit/i }).click();
-    await page.waitForURL(/\/household\//, { timeout: 45000 });
+    await fillPinInput(page, MEMBER_PIN);
+    await page.waitForURL(/\/household\//, { timeout: 30000 });
   });
 
   test('shows Household Overview section', async ({ page }) => {
